fix(cli): parse region flag (#919)
https://coveord.atlassian.net/browse/CDX-1139

<!-- For Coveo Employees only. Fill this section.

CDX-1139

-->

## Proposed changes

In the OAuth library, we compare the region using the 'value' of the
enum, and in the flag we use the 'key' of the enum.
This causes the comparison to fail (`'US' !== 'us'`).

So to fix that, I sanitize the input and ensure everything is lowercase
only. I also improved the region flag to accept both the lower and upper
case variations: it can be a pain to have to go upper case just for
typing a region.

## Testing

- [x] Unit Tests:
<!-- Did you write unit tests for your feature? If not, explains why?
-->
- [ ] Functionnal Tests:
<!-- Did you write functionnal tests for your feature? If not, explains
why? -->
- [x] Manual Tests: tested with /bin/dev:
- Run `bin/dev auth:login -r=us -e=hipaa`: it fails before, now it works
- Run `bin/dev auth:login -r=US -e=hipaa`: before, the OAuth redirection
was dropped, now it works.