chore(npm): do not npmignore angular/dist + unlock npm (#805)
https://coveord.atlassian.net/browse/CDX-1010

<!-- For Coveo Employees only. Fill this section.

CDX-1010

-->

## Proposed changes

I discussed with a developer from the npm team here
https://github.com/npm/npm-packlist/issues/106 and got a good
understanding of their philosophy moving forward with the concept of
workspaces and their multi-level ignore files.

In short, npm will try to acquire an ignorefile at each directory level
and concatenate them. The peculiarity (at least IMO), is that if it
acquire a npmignore first, this will have no incidence for the
acquisition of the other ignorefile. So, we can end up with a mix of
`gitignore` and `npmignore`. I discussed this behavior with the npm
team, and they are kinda stuck, so its their least worst option here, so
we just have to adapt and make it work.

This is what this PR is about.

## Testing

 - Ran `npm publish --dry-run` on every package and checked them out.
 
 ## Prerequisites
 
 - [x] https://github.com/npm/npm-packlist/pull/108
 - [ ] `npm-cli` released with ☝️