chore: migrate to independent release (#929)
<!-- For Coveo Employees only. Fill this section.

CDX-1146

-->

## Proposed changes

Modify the release process to be able to do independent releases.
This also means that package version will not be 'synchronized' anymore:
So in the future, the default version of `create-atomic` for the `cli`
might not be the version of the `cli`.

## Breaking changes

### How we do merge for changes that impact multiple projects:
- For now: If your changes impact multiple packages, you have to
organize your commits so that one commit affects a single workspace
every time.
- Soon ™️ : The PR description will be changed to include a new section
that will be used by semantic release parser to make 'one commit nth
commits'. This will be done in another PR

For example, if I do a breaking change in `@coveo/cli-commons`, I will
need to change how `@coveo/cli` use the latter.
I will need to organize my branch in two commits:
- The first one will contain only the change relative to
`@coveo/cli-commons`. Its message will contain the breaking change
identifier.
- The second one will contain only the change relative to `@coveo/cli`.
Its message will not contain the breaking change container, it will be a
chore or feat depending on the impact on the user. (e.g. if the change
is not impacting the end user of the CLI, then it's a chore. If the
change adds new behavior for the end user, it's a feat)

For this reason, I enabled the rebase merge option on the repository. It
will be the one to use for the scenario described above.
### Tags format

- A tag for each release process, named `release-n`. `release-0` has
been placed manually on the latest release.
- A tag for each released package, named `@coveo/package-name@version`.

### GitHub release
- Will occur only when the CLI is released.

## Testing

- [x] Manual Tests: I created a test branch locally where I disabled git
push and npm publish and tested the process.