test(e2e): uplink all packages then publish (#954)
<!-- For Coveo Employees only. Fill this section.

CDX-1162

-->

## Proposed changes

### Uplink first, publish changed packages after that
I thought that 'any' version of the package published prior would
satisfy `npm version`, but no it need exactly the previous one.

To allow this, I changed how the scaffolding work: it now do what uplink
was doing: copying all the existing packages into verdaccio.
This may seems counter-intuitive and one might thing that we're not
testing our code but the publish packages, but that would not be the
case, because only the unchanged packages will be untouched. The other
will have a new package version published

### Ensure we published packages that are changed 'transitively'

Let's say I have a PR or release that only affects `@coveo/cli-commons`,
in the current state, `@coveo/cli`would not be released/published on
verdaccio. This changes that by checking if the `package.json`has been
modified if there's been no commit on the package directory