docs(cli): when to use section (#974)
<!-- For Coveo Employees only. Fill this section.

CDX-1204

-->

## Proposed changes

As I was looking at the [google
doc](https://docs.google.com/document/d/1s6m31vbUwGk8uv87wg9bAOOa3J2KzLYtID_DlHrXBJ0/edit),
I thought that it could fit well in the main readme, since it's useful
material to cover in an introduction.
In addition to appearing in this GitHub repo, it will also be fetched
automatically by a script so it appears in this page of the docs site:
https://docs.coveo.com/en/cli/

While I was there, I also made minor edits in the preceding section.