chore(prerelease): check beta tag and not latest on prerelease (#1019)
<!-- For Coveo Employees only. Fill this section.
[CDX-1223](https://coveord.atlassian.net/browse/CDX-1223)
-->

## Proposed changes

After publishing, we were using `npm show` to validate that the package
was released properly... but we were doing so on the `latest` tag.
This cannot work when we work with `beta` tag.

So I changed that. And then I also DRY: `npm show` is essentially what
`describeNpmTag` was doing: less code, less bug (maybe).