chore(prerelease): do meaningful prerelease (#1025)
<!-- For Coveo Employees only. Fill this section.

CDX-1216

-->

## Proposed changes

While working, due to the misfire that occurred, the prerelease version
is not really meaningful: the behaviour I wanted was 'the next release
prefixed by -someNumber'. Now it is not, and if other mishaps happened,
it could have been the same.

So to fix that, I changed how we define the beta version: we base
strictly ourselves on two things:
 - the next 'gold/release' version.
 - the versions existing in the registry.

If we find no unprefixed prerelease for the next gold, then we create
the -0 one, otherwise, we took the latest and 'bump it'

## Testing:

On d8c5c7a252fd5268b0bbedaec6d25a11cbd88662, run `npm run release:phase1
-w=@coveo/cli`.
Prior to the change, it was doing `3.0.0-1`. Now it is doing `2.0.0-1`.