feat(cli): add fallback port for oauth (#1063)
<!-- For Coveo Employees only. Fill this section.

CDX-1252

-->

Fixes #1061 

## Proposed changes

Add 52296 as an alternative port for OAuth. I choose 52296 because it is
part of the private port range as defined by IANA, so I think the
collision chance is lesser than in the system or user range. A caveat,
this could be argued as a misuse because the port cannot be 'dynamically
assigned' (i.e. the program can't change it on a whim). But I reckon the
benefits outweigh this technicality.

## Testing

- [x] Unit Tests:
- [x] Manual Tests: reproducing steps of #1061