chore(workflows): don't use always() in workflows (#1274)
Related Issue
DT-5692

## Hi, what's this?
👋 We have come across a problem with using `if: always()` in workflows.
The basic gist is that these steps become un-cancellable, possibly
creating all sorts of unexpected problems. This is less frequent with
GitHub hosted runners, but it can happen.

This PR introduces and workaround this problem by explicitly defining
the different states the step can face.
Namely:
* `success()`
* `cancelled()`
* `failure()`

This way, you get the `always()` behaviour without it's possible
caveats.

## I don't want this change
In a way, this is more of a "we really recommend this" kind of PR. That
is because you may never face the problem I'm talking about, but when
you do, you'll want this workaround.

If you decide against merging this, please make sure you mention this if
you need support in the future to help us identify this issue faster.

## Final notes
Please review and merge this PR. Since this is a CI change, you don't
need to release this change.

This PR is created as part of a large batch. I would appreciate it if
you could merge this yourself to ease the amount of followup on my end.

## References
* https://coveo.slack.com/archives/CBP3GTFTQ/p1679495407426389 (Coveo
internal)
* https://github.com/orgs/community/discussions/26303