test: add missing proxied packages (#1422)
When running our E2E, we publish all the packages of the repo into a
temp Verdaccio registry.
To do so nicely, we configured Verdaccio in a way that essentially goes
by "If it's a package in this repo, I can handle it, if not, I'll ring
NPM"

Well err, we forget to add a few packages to the list. This fixes that.