feat(resources): support catalogs in resource snapshots (#1499)
<!-- For Coveo Employees only. Fill this section.

SRC-6089
-->

## Proposed changes

Need to update platform-client to support new resource types CATALOG and
CATALOG_CONFIG.

<!--
Explains what are you changing in the code and provide links to relevant
issues.
This doesn't need to be over-technical, a good summary of the issues
with good explanations of the decisions should be enough.
 -->

## Breaking changes

In new platform-client version, CreateApiKeyModel does not have
`enabled` attribute.

<!--
    Remove this section if the PR does not include any breaking change

If your changes includes some breaking changes in the code, thoroughly
explains:
        - What are the breaking changes programmatically speaking.
- What is the impact on the end-user (e.g. user cannot do X anymore).
        - What motivates those changes.
-->

## Testing

- [ ] Unit Tests:
<!-- Did you write unit tests for your feature? If not, explains why?
-->
- [ ] Functionnal Tests:
<!-- Did you write functionnal tests for your feature? If not, explains
why? -->
- [ ] Manual Tests:
<!-- How did you test your changeset?  -->
I created a snapshot and applied it to another org with catalogs.